+++ sdk/identity/identity-common/src/credentials/vsCodeCredentialControl.ts
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { VSCodeCredentialFinder } from "./visualStudioCodeCredentialPlugin";
+
+let findCredentials: VSCodeCredentialFinder | undefined = undefined;
+
+export const vsCodeCredentialControl = {
+  setVsCodeCredentialFinder(finder: VSCodeCredentialFinder): void {
+    findCredentials = finder;
+  },
+  getVsCodeCredentialFinder(): VSCodeCredentialFinder| undefined {
+    return findCredentials
+  }
+};
