+++ sdk/identity/identity-common/src/plugins/provider.ts
@@ -26,6 +26,7 @@ export interface CachePluginControl {
  */
 export interface VisualStudioCodeCredentialControl {
   setVsCodeCredentialFinder(finder: VSCodeCredentialFinder): void;
+  getVsCodeCredentialFinder(): VSCodeCredentialFinder|undefined;
 }
 
 /**
