+++ sdk/identity/identity/src/index.ts
@@ -1,12 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export * from "./plugins/consumer";
-
-export { IdentityPlugin } from "./plugins/provider";
-
 import { TokenCredential } from "@azure/core-auth";
 import { DefaultAzureCredential } from "./credentials/defaultAzureCredential";
+import  { IdentityPlugin } from "../../identity-common/src/plugins/provider";
+import { useIdentityPlugin } from "../../identity-common/src/plugins/consumer";
+export { IdentityPlugin, useIdentityPlugin };
 
 export {
   AuthenticationError,
